+++ src/main/java/carGame/LengthLimitException.java
@@ -1,0 +1,8 @@
+package carGame;
+
+public class LengthLimitException extends Exception {
+
+    public LengthLimitException(String message) {
+        super(message);
+    }
+}
